requirements.txt
pandas>=0.25.1
seaborn>=0.9.0
alpaca-ml==0.8.0
torch>=1.2.0
torchvision=
catalyst==20.3
matplotlib>=3.0.0
albumentations==0.4.3

